requirements.txt
cnocr==2.3.1
colormath==3.0.0
mss==10.0.0
numpy==2.2.3
opencv_contrib_python==4.11.0.86
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
paddleocr==2.9.1
paddleocr.egg==info
pefile==2024.8.26
Pillow==11.1.0
PyAutoGUI==0.9.54
pynput==1.7.7
PyQt6==6.8.1
PyQt6_Fluent_Widgets==1.7.5
PyQt6_sip==13.10.0
PyQt_Fluent_Widgets==1.7.5
pytesseract==0.3.13
pywin32==308
